Process record byte array to String
This changes the plugin from using a character set-specific decoder to
simply parsing each record's byte array to a String. This punts the
character set responsibility downstream to the codec plugin used in
conjunction with this input. This also opens up the potential use of
alternative codecs like gzip_lines.