Fix problem in loggers config when run with Logstah >= 7.9
This plugin customized the log level for a logger assuming the logger implementation was Jdk14Logger and logger framework Apache Commons Logging.
But starting with Logstah 7.9 when was added log4j-jcl bridge created interference and changed the logger impl to Log4J JCL.
This commit adapt to both cases.

Fix 87